More testing. Moved models to their own directory...
Having everything top level is all ... icky. They should be somewhat
better organized.

While there isn't technically a web app yet, it can now be web-iffied,
which is cool. What I mean is that all of the tests are passing, and
we've got words, pages, and other such things in the db. Which is cool.

Also, while not used yet, there's a `PageLinks` model, which just knows
who links to who, which is also cool.